+++ kernel/src/drivers/GPUsupport/telemetry/counters.rs
@@ -2,13 +2,13 @@
  * GPU counters.
  */
 
-use core::sync::atomic::{AtomicU64, Ordering};
+use core::sync::atomic::{AtomicU32, Ordering};
 
-pub static GPU_FAULT_COUNT: AtomicU64 = AtomicU64::new(0);
-pub static GPU_FENCE_STALLS: AtomicU64 = AtomicU64::new(0);
-pub static GPU_QUEUE_SUBMITS: AtomicU64 = AtomicU64::new(0);
+pub static GPU_FAULT_COUNT: AtomicU32 = AtomicU32::new(0);
+pub static GPU_FENCE_STALLS: AtomicU32 = AtomicU32::new(0);
+pub static GPU_QUEUE_SUBMITS: AtomicU32 = AtomicU32::new(0);
 
-pub fn snapshot() -> (u64, u64, u64) {
+pub fn snapshot() -> (u32, u32, u32) {
     (
         GPU_FAULT_COUNT.load(Ordering::Relaxed),
         GPU_FENCE_STALLS.load(Ordering::Relaxed),
